abstraction made from CHSH versions to CHSH abstract class CHSH.py
