--- CHSHgame/rewardFunctions.py
@@ -1,2 +1,0 @@
-def reward_only_best():
-    ...
